Refactored into separate foldiers and subpackages
Moved Connection util out of service.
