+++ project0/src/main/java/com/revature/anthony/nguyen/project0/project0/dao/AccountCheckingDao.java
@@ -1,4 +1,4 @@
-package com.revature.anthony.nguyen.project0.project0;
+package com.revature.anthony.nguyen.project0.project0.dao;
 
 import java.util.Optional;
